Simplify hypertable DDL APIs
The current implementation of hypertable DDL APIs is very much inclined towards
tables having a "time" column. The idea with this patch is to implement generic
hypertable DDL APIs which can be used to convert a regular PostgreSQL table to
hypertable irrespective of the data type of partition column.

The new APIs are being added to "timescaledb_experimental" schema and the
transition from old to new APIs is based on the following factors:
- Remove/update attributes having specific time reference
- Drop attributes related to deprecated features
- Remove attributes with minimal usage patterns
- Remove attributes specific to distributed hypertable

Similar changes are done in DDL API code path to remove "time" specific
references. The APIs affected as part of these changes are:
- create_hypertable
- add_dimension
- set_chunk_time_interval

Also, added support for using SERIAL/BIGSERIAL as partition columns.